scale

+++ MNIST/LeNet-5/kaggle_mnist_train.py
@@ -84,7 +84,7 @@ def main(argv=None):
     #(test_images,num_test,num_feature)=kaggle_mnist_input_data.read_test_data(test_filename)
     #将DataFrame转化为Matrix
     training_images=train_images.as_matrix().astype('float64')/255.0
-    training_labels=train_labels.as_matrix().astype('float64')
+    training_labels=train_labels.as_matrix()
     #testing_images=test_images.as_matrix()
     training_labels_onehot=kaggle_mnist_input_data.dense_to_one_hot(training_labels,num_classes=NUM_CLASS)
